Fixed #31
I rewrote chunks of the HTML to get it more in-line with other pages. I tweaked some of the CSS, and moved the important bits into main.css. Redundant styles were removed, and so about.css was ultimately removed.

I added a secondary navigation (nav.secondary) so I had to also rename the main navigation (nav.main). This had to be changed on every other page...