Began working on avatars
Avatars are now pulled from /media/avatars/ and they have a filename of [user's email address].png . Obviously this still needs some work, because not all avatars will be PNG. Perhaps all can be resized on upload and stored as JPG?

I added User::avatar() function that returns an appropriate URL to use for an avatar. To use, you must pass in an associative array with an 'email' field OR you must pass in an email address. This too may need some tweaking...