FAVES now work
This uses AJAX similar to how LIKES work. I did notice a strange bug, if the number of faves in the database falls to 0, MySQL reports that there are no likes as well. If the error occurs on 0-faves-per-project (instead of the total in the DB), then we have a bug that definitely needs fixing. It probably has to do with how the JOIN is occurring in the MySQL query.